fix: add __main__.py, extract _wizard.py and _main.py from cli package
