+++ factory/cli/__main__.py
@@ -1,0 +1,3 @@
+from factory.cli import main
+
+raise SystemExit(main())
